backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai
markdown
flask-sqlalchemy
flask-login
werkzeug
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0 